backend/requirements.txt
flask
flask-cors
google-generativeai
langdetect
python-dotenv
requests
gunicorn
pytest
pypdf
PyPDF2
pandas
openpyxl
python-docx
python-pptx
Pillow
pytesseract
opencv-python
numpy
xlrd
